+++ src/main/java/org/cojen/tupl/BoundedCursor.java
@@ -47,8 +47,8 @@ final class BoundedCursor implements Cursor {
     }
 
     @Override
-    public void setValueLength(long length) throws IOException {
-        mSource.setValueLength(length);
+    public void valueLength(long length) throws IOException {
+        mSource.valueLength(length);
     }
 
     @Override