+++ src/main/java/org/cojen/tupl/Tree.java
@@ -164,7 +164,7 @@ class Tree implements View, Index {
         try {
             if (highKey != null) {
                 high = newCursor(Transaction.BOGUS);
-                high.autoload(false);
+                high.mKeyOnly = true;
                 high.find(highKey);
                 if (high.mKey == null) {
                     // Found nothing.
@@ -501,7 +501,7 @@ class Tree implements View, Index {
         keyCheck(key);
         TreeCursor cursor = newCursor(txn);
         try {
-            cursor.autoload(false);
+            cursor.mKeyOnly = true;
             cursor.findAndStore(key, value);
         } finally {
             cursor.reset();
@@ -524,7 +524,7 @@ class Tree implements View, Index {
         keyCheck(key);
         TreeCursor cursor = newCursor(txn);
         try {
-            cursor.autoload(false);
+            cursor.mKeyOnly = true;
             return cursor.findAndModify(key, TreeCursor.MODIFY_INSERT, value);
         } finally {
             cursor.reset();
@@ -536,7 +536,7 @@ class Tree implements View, Index {
         keyCheck(key);
         TreeCursor cursor = newCursor(txn);
         try {
-            cursor.autoload(false);
+            cursor.mKeyOnly = true;
             return cursor.findAndModify(key, TreeCursor.MODIFY_REPLACE, value);
         } finally {
             cursor.reset();
@@ -669,6 +669,11 @@ class Tree implements View, Index {
         return isClosed();
     }
 
+    @Override
+    public final boolean isModifyAtomic() {
+        return true;
+    }
+
     /**
      * Current approach for evicting data is as follows:
      * - Search for a random Node, steered towards un-cached nodes. 
@@ -752,7 +757,7 @@ class Tree implements View, Index {
     public Stats analyze(byte[] lowKey, byte[] highKey) throws IOException {
         TreeCursor cursor = newCursor(Transaction.BOGUS);
         try {
-            cursor.autoload(false);
+            cursor.mKeyOnly = true;
             cursor.random(lowKey, highKey);
             return cursor.key() == null ? new Stats(0, 0, 0, 0, 0) : cursor.analyze();
         } catch (Throwable e) {
@@ -787,7 +792,7 @@ class Tree implements View, Index {
 
         TreeCursor cursor = newCursor(Transaction.BOGUS);
         try {
-            cursor.autoload(false);
+            cursor.mKeyOnly = true;
 
             // Find the first node instead of calling first() to ensure that cursor is
             // positioned. Otherwise, empty trees would be skipped even when the root node
@@ -833,7 +838,7 @@ class Tree implements View, Index {
     final boolean verifyTree(Index view, VerificationObserver observer) throws IOException {
         TreeCursor cursor = newCursor(Transaction.BOGUS);
         try {
-            cursor.autoload(false);
+            cursor.mKeyOnly = true;
             cursor.first(); // must start with loaded key
             int height = cursor.height();
             if (!observer.indexBegin(view, height)) {
@@ -1411,10 +1416,10 @@ class Tree implements View, Index {
 
         void prime() {
             try {
-                Cursor c = newCursor(Transaction.BOGUS);
+                TreeCursor c = newCursor(Transaction.BOGUS);
 
                 try {
-                    c.autoload(false);
+                    c.mKeyOnly = true;
 
                     while (true) {
                         byte[] key;