+++ src/main/java/org/cojen/tupl/repl/Controller.java
@@ -51,7 +51,10 @@ import java.util.logging.Level;
 import org.cojen.tupl.util.Latch;
 import org.cojen.tupl.util.LatchCondition;
 
-import static org.cojen.tupl.io.Utils.*;
+import org.cojen.tupl.io.Utils;
+import static org.cojen.tupl.io.Utils.closeQuietly;
+import static org.cojen.tupl.io.Utils.encodeLongLE;
+import static org.cojen.tupl.io.Utils.rethrow;
 
 /**
  * Core replicator implementation.
@@ -69,7 +72,8 @@ final class Controller extends Latch implements StreamReplicator, Channel {
     private static final int JOIN_TIMEOUT_MILLIS = 5000;
     private static final int MISSING_DATA_REQUEST_SIZE = 100_000;
 
-    private static final byte CONTROL_OP_JOIN = 1, CONTROL_OP_UPDATE_ROLE = 2;
+    private static final byte CONTROL_OP_JOIN = 1, CONTROL_OP_UPDATE_ROLE = 2,
+        CONTROL_OP_UNJOIN = 3;
 
     private static final byte[] EMPTY_DATA = new byte[0];
 
@@ -517,9 +521,12 @@ final class Controller extends Latch implements StreamReplicator, Channel {
                     newRole = mGroupFile.localMemberRole();
                 }
                 break;
+            case CONTROL_OP_UNJOIN:
+                refresh = mGroupFile.applyRemovePeer(message);
+                break;
             }
 
-            // FIXME: Followers should inform the leader very early of their current group
+            // TODO: Followers should inform the leader very early of their current group
             // version, and if they inform the leader again whenever it changes. This speeds up
             // updateRole without forcing the caller to retry.
 
@@ -738,7 +745,7 @@ final class Controller extends Latch implements StreamReplicator, Channel {
                 commitIndex = writer.mCommitIndex;
             }
 
-            // FIXME: stream it
+            // TODO: stream it
             data = Arrays.copyOfRange(data, offset, offset + length);
 
             for (Channel peerChan : peerChannels) {
@@ -782,6 +789,12 @@ final class Controller extends Latch implements StreamReplicator, Channel {
         mChanMan.stop();
         mScheduler.shutdown();
         mStateLog.close();
+    }
+
+    void uncaught(Throwable e) {
+        if (!mChanMan.isStopped()) {
+            Utils.uncaught(e);
+        }
     }
 
     Scheduler scheduler() {
@@ -857,7 +870,7 @@ final class Controller extends Latch implements StreamReplicator, Channel {
     private void requestMissingData(final long startIndex, final long endIndex) {
         event(Level.FINE, () -> "Requesting missing data: [" + startIndex + ", " + endIndex + ')');
 
-        // FIXME: Need a way to abort outstanding requests.
+        // TODO: Need a way to abort outstanding requests.
 
         long remaining = endIndex - startIndex;
         long index = startIndex;
@@ -951,7 +964,7 @@ final class Controller extends Latch implements StreamReplicator, Channel {
             }
 
             // Convert to candidate.
-            // FIXME: Don't permit rogue members to steal leadership if process is
+            // TODO: Don't permit rogue members to steal leadership if process is
             // suspended. Need to double check against local clock to detect this. Or perhaps
             // check with peers to see if leader is up.
             mLocalMode = MODE_CANDIDATE;
@@ -1088,8 +1101,14 @@ final class Controller extends Latch implements StreamReplicator, Channel {
 
     private void requestJoin(Socket s) {
         try {
-            if (doRequestJoin(s)) {
-                return;
+            try {
+                if (doRequestJoin(s)) {
+                    return;
+                }
+            } catch (IllegalStateException e) {
+                // Cannot remove local member.
+                OutputStream out = s.getOutputStream();
+                out.write(new byte[] {GroupJoiner.OP_ERROR, ErrorCodes.INVALID_ADDRESS});
             }
         } catch (Throwable e) {
             closeQuietly(s);
@@ -1105,13 +1124,24 @@ final class Controller extends Latch implements StreamReplicator, Channel {
     private boolean doRequestJoin(Socket s) throws IOException {
         ChannelInputStream in = new ChannelInputStream(s.getInputStream(), 100);
 
+        SocketAddress addr;
+        long memberId;
+
         int op = in.read();
 
-        if (op != GroupJoiner.OP_ADDRESS) {
+        switch (op) {
+        default:
             return joinFailure(s, ErrorCodes.UNKNOWN_OPERATION);
+        case GroupJoiner.OP_ADDRESS:
+        case GroupJoiner.OP_UNJOIN_ADDRESS:
+            addr = GroupFile.parseSocketAddress(in.readStr(in.readIntLE()));
+            memberId = 0;
+            break;
+        case GroupJoiner.OP_UNJOIN_MEMBER:
+            addr = null;
+            memberId = in.readLongLE();
+            break;
         }
-
-        SocketAddress addr = GroupFile.parseSocketAddress(in.readStr(in.readIntLE()));
 
         OutputStream out = s.getOutputStream();
 
@@ -1138,52 +1168,102 @@ final class Controller extends Latch implements StreamReplicator, Channel {
             return joinFailure(s, ErrorCodes.NO_ACCEPTOR);
         }
 
-        byte[] message = mGroupFile.proposeJoin(CONTROL_OP_JOIN, addr, (gfIn, index) -> {
-            // Join is accepted, so reply with the log index info and the group file. The index
-            // is immediately after the control message.
+        byte[] message;
 
-            try {
-                if (gfIn == null) {
-                    out.write(new byte[] {
-                        // Assume failure is due to version mismatch, but could be something else.
-                        // TODO: Use more specific error code.
-                        GroupJoiner.OP_ERROR, ErrorCodes.VERSION_MISMATCH});
-                    return;
-                }
+        switch (op) {
+        case GroupJoiner.OP_ADDRESS:
+            message = mGroupFile.proposeJoin(CONTROL_OP_JOIN, addr, (gfIn, index) -> {
+                // Join is accepted, so reply with the log index info and the group file. The
+                // index is immediately after the control message.
 
-                TermLog termLog = mStateLog.termLogAt(index);
-
-                byte[] buf = new byte[1000];
-                int off = 0;
-                buf[off++] = GroupJoiner.OP_JOINED;
-                encodeLongLE(buf, off, termLog.prevTermAt(index)); off += 8;
-                encodeLongLE(buf, off, termLog.term()); off += 8;
-                encodeLongLE(buf, off, index); off += 8;
-
-                while (true) {
-                    int amt;
-                    try {
-                        amt = gfIn.read(buf, off, buf.length - off);
-                    } catch (IOException e) {
-                        // Report any GroupFile corruption.
-                        uncaught(e);
+                try {
+                    if (gfIn == null) {
+                        out.write(new byte[] {
+                            // Assume failure is due to version mismatch, but could be
+                            // something else.
+                            // TODO: Use more specific error code.
+                            GroupJoiner.OP_ERROR, ErrorCodes.VERSION_MISMATCH
+                        });
                         return;
                     }
-                    if (amt < 0) {
+
+                    TermLog termLog = mStateLog.termLogAt(index);
+
+                    byte[] buf = new byte[1000];
+                    int off = 0;
+                    buf[off++] = GroupJoiner.OP_JOINED;
+                    encodeLongLE(buf, off, termLog.prevTermAt(index)); off += 8;
+                    encodeLongLE(buf, off, termLog.term()); off += 8;
+                    encodeLongLE(buf, off, index); off += 8;
+
+                    while (true) {
+                        int amt;
+                        try {
+                            amt = gfIn.read(buf, off, buf.length - off);
+                        } catch (IOException e) {
+                            // Report any GroupFile corruption.
+                            uncaught(e);
+                            return;
+                        }
+                        if (amt < 0) {
+                            break;
+                        }
+                        out.write(buf, 0, off + amt);
+                        off = 0;
+                    }
+                } catch (IOException e) {
+                    // Ignore.
+                } finally {
+                    closeQuietly(out);
+                }
+            });
+
+            break;
+
+        case GroupJoiner.OP_UNJOIN_ADDRESS:
+            // Find the member and fall through to the next case.
+            if (mGroupFile.localMemberAddress().equals(addr)) {
+                memberId = mGroupFile.localMemberId();
+            } else {
+                for (Peer peer : mGroupFile.allPeers()) {
+                    if (peer.mAddress.equals(addr)) {
+                        memberId = peer.mMemberId;
                         break;
                     }
-                    out.write(buf, 0, off + amt);
-                    off = 0;
                 }
-            } catch (IOException e) {
-                // Ignore.
-            } finally {
-                closeQuietly(out);
             }
-        });
+
+        case GroupJoiner.OP_UNJOIN_MEMBER:
+            message = mGroupFile.proposeRemovePeer(CONTROL_OP_UNJOIN, memberId, success -> {
+                try {
+                    byte[] reply;
+
+                    if (success) {
+                        reply = new byte[] {GroupJoiner.OP_UNJOINED};
+                    } else {
+                        reply = new byte[] {
+                            // Assume failure is due to version mismatch, but could be
+                            // something else.
+                            // TODO: Use more specific error code.
+                            GroupJoiner.OP_ERROR, ErrorCodes.VERSION_MISMATCH
+                        };
+                    }
+
+                    out.write(reply);
+                } catch (IOException e) {
+                    // Ignore.
+                } finally {
+                    closeQuietly(out);
+                }
+            });
+            break;
+
+        default:
+            throw new AssertionError();
+        }
 
         mScheduler.schedule(() -> {
-            if (mGroupFile.discardJoinConsumer(message)) {
+            if (mGroupFile.discardProposeConsumer(message)) {
                 closeQuietly(s);
             }
         }, JOIN_TIMEOUT_MILLIS);
@@ -1467,7 +1547,7 @@ final class Controller extends Latch implements StreamReplicator, Channel {
             return true;
         }
 
-        // FIXME: Gracefully handle stale data requests, received after log compaction.
+        // TODO: Gracefully handle stale data requests, received after log compaction.
 
         try {
             LogReader reader = mStateLog.openReader(startIndex);
@@ -1511,7 +1591,7 @@ final class Controller extends Latch implements StreamReplicator, Channel {
                         break;
                     }
 
-                    // FIXME: stream it
+                    // TODO: stream it
                     byte[] data = new byte[amt];
                     System.arraycopy(buf, 0, data, 0, amt);
 
@@ -1574,7 +1654,7 @@ final class Controller extends Latch implements StreamReplicator, Channel {
             LogWriter writer = mStateLog.openWriter(prevTerm, term, index);
 
             if (writer == null) {
-                // FIXME: stash the write for later (unless it's a stale write)
+                // TODO: stash the write for later (unless it's a stale write)
                 // Cannot write because terms are missing, so request them.
                 long now = System.currentTimeMillis();
                 if (now >= mNextQueryTermTime) {
@@ -1867,7 +1947,7 @@ final class Controller extends Latch implements StreamReplicator, Channel {
 
             if (peer == null || peer.mMemberId != memberId) {
                 // Member doesn't exist.
-                // FIXME: Permit leader to change itself.
+                // TODO: Permit leader to change itself.
                 result = ErrorCodes.UNKNOWN_MEMBER;
                 break tryUpdateRole;
             }