+++ src/test/java/org/cojen/tupl/DefaultCursor.java
@@ -39,8 +39,8 @@ class DefaultCursor implements Cursor {
     }
 
     @Override
-    public void setValueLength(long length) throws IOException {
-        mSource.setValueLength(length);
+    public void valueLength(long length) throws IOException {
+        mSource.valueLength(length);
     }
 
     @Override